#225 [FEAT] : 0.5초 이하 이벤트 처리 클릭 리스너 확장 함수 추가
